requirements.txt
Django==2.1
Pillow
urllib3
certifi
python-Levenshtein
